chore: optimize performance by reducing definition updates
This commits aims to reduce expensive web client updates by only
triggering them when needed. This is achieved in two ways:

1. The `Feedback` class implementation requires a new method,
`selectCheckFeedback()`, which returns an `Observable` that will only
emit when the feedback needs to be updated. This allows using the
`selectRefresh()` method exclusively for definitions update events,
that is, when the dropdown menus need to be updated.

2. A new `filterEntitiesChanged()` RXJS operator, to filter events and
only emit when either the list length changed, or the entities names
changed.